Refactor to new endpoint
Change endpoint to be `/inspect/graph/patient/id` instead of `/inspect/patient/id`. This is to deconflict with Lakshmi's timeline visualisation.
